requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-google-genai
faiss-cpu
PyPDF2
